Extracted common Scheduler code to shared directory and made methods on Scheduler instance private[fs2]
