+++ core/jvm/src/main/scala/fs2/SchedulerPlatform.scala
@@ -2,39 +2,13 @@ package fs2
 
 import java.util.concurrent.{ Executors, ScheduledExecutorService, TimeUnit }
 
-import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.FiniteDuration
 
 import cats.effect.Sync
 
 import fs2.internal.ThreadFactories
 
-/** Provides the ability to schedule evaluation of thunks in the future. */
-trait Scheduler {
-
-  /**
-   * Evaluates the thunk after the delay.
-   * Returns a thunk that when evaluated, cancels the execution.
-   */
-  def scheduleOnce(delay: FiniteDuration)(thunk: => Unit): () => Unit
-
-  /**
-   * Evaluates the thunk every period.
-   * Returns a thunk that when evaluated, cancels the execution.
-   */
-  def scheduleAtFixedRate(period: FiniteDuration)(thunk: => Unit): () => Unit
-
-  /**
-   * Returns an execution context that executes all tasks after a specified delay.
-   */
-  def delayedExecutionContext(delay: FiniteDuration, reporter: Throwable => Unit = ExecutionContext.defaultReporter): ExecutionContext = new ExecutionContext {
-    def execute(runnable: Runnable): Unit = { scheduleOnce(delay)(runnable.run); () }
-    def reportFailure(cause: Throwable): Unit = reporter(cause)
-    override def toString = s"DelayedExecutionContext($delay)"
-  }
-}
-
-object Scheduler {
+private[fs2] trait SchedulerPlatform {
   /** Returns a singleton stream consisting of a `Scheduler`. The scheduler is shutdown at finalization of the stream. */
   def apply[F[_]](corePoolSize: Int, daemon: Boolean = true, threadPrefix: String = "fs2-scheduler", exitJvmOnFatalError: Boolean = true)(implicit F: Sync[F]): Stream[F,Scheduler] =
     Stream.bracket(allocate(corePoolSize, daemon, threadPrefix, exitJvmOnFatalError))(t => Stream.emit(t._1), _._2)
