+++ core/shared/src/main/scala/fs2/Scheduler.scala
@@ -2,22 +2,21 @@ package fs2
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.FiniteDuration
-import scala.scalajs.js.timers._
 
 /** Provides the ability to schedule evaluation of thunks in the future. */
-trait Scheduler {
+abstract class Scheduler {
 
   /**
    * Evaluates the thunk after the delay.
    * Returns a thunk that when evaluated, cancels the execution.
    */
-  def scheduleOnce(delay: FiniteDuration)(thunk: => Unit): () => Unit
+  private[fs2] def scheduleOnce(delay: FiniteDuration)(thunk: => Unit): () => Unit
 
   /**
    * Evaluates the thunk every period.
    * Returns a thunk that when evaluated, cancels the execution.
    */
-  def scheduleAtFixedRate(period: FiniteDuration)(thunk: => Unit): () => Unit
+  private[fs2] def scheduleAtFixedRate(period: FiniteDuration)(thunk: => Unit): () => Unit
 
   /**
    * Returns an execution context that executes all tasks after a specified delay.
@@ -29,16 +28,4 @@ trait Scheduler {
   }
 }
 
-object Scheduler {
-  val default: Scheduler = new Scheduler {
-    override def scheduleOnce(delay: FiniteDuration)(thunk: => Unit): () => Unit = {
-      val handle = setTimeout(delay)(thunk)
-      () => { clearTimeout(handle) }
-    }
-    override def scheduleAtFixedRate(period: FiniteDuration)(thunk: => Unit): () => Unit = {
-      val handle = setInterval(period)(thunk)
-      () => { clearInterval(handle) }
-    }
-    override def toString = "Scheduler"
-  }
-}
+object Scheduler extends SchedulerPlatform
